requirements.txt
# Core
requests>=2.31.0
httpx>=0.27.0
pandas>=2.2.0
python-dotenv>=1.0.0

# Fuzzy matching
rapidFuzz>=3.6.0

# Google Sheets
gspread>=6.0.0
google-auth>=2.28.0

# CLI
typer>=0.12.0

# Scraping (optional, for Playwright path)
# playwright>=1.43.0

# Outscraper (optional)
# outscraper>=4.3.0
